agent/requirements.txt
groq
mysql-connector-python
python-dotenv
